refactor section nav to smaller modules
