+++ spotlight-client/src/SystemNarrativePage/SystemNarrativePage.tsx
@@ -30,7 +30,7 @@ import normalizeRouteParams from "../routerUtils/normalizeRouteParams";
 import { colors, typefaces, Chevron } from "../UiLibrary";
 import { X_PADDING } from "./constants";
 import Section from "./Section";
-import SectionNavigation from "./SectionNavigation";
+import NarrativeNavigation from "../NarrativeNavigation";
 
 const Container = styled.article`
   display: flex;
@@ -171,7 +171,7 @@ const SystemNarrativePage: React.FC<{
       <NavContainer>
         <Sticker>
           <NavStickyContainer>
-            <SectionNavigation
+            <NarrativeNavigation
               activeSection={activeSection}
               narrative={narrative}
             />
